Divide tasklist class into smaller classes
