+++ data/tasks.txt
@@ -1,0 +1,4 @@
+T | 0 | haha
+T | 0 | me /to lol
+E | 0 | frrom | 10-10-2005 12:30 | 10-11-2005 12:30
+T | 0 | lol
