+++ src/main/java/darkpool/Darkpool.java
@@ -5,24 +5,14 @@ import darkpool.gui.Gui;
 import darkpool.util.DarkpoolException;
 import darkpool.parser.Parser;
 import darkpool.util.Storage;
-import darkpool.util.TaskList;
+import darkpool.tasklist.TaskList;
 
-/**
- * The Darkpool class represents the main logic of the application.
- * It handles the initialization of the storage, task list, and GUI components.
- * It also processes user input and returns appropriate responses.
- */
 public class Darkpool {
 
     private final Storage storage;
     private TaskList taskList;
     private final Gui gui;
 
-    /**
-     * Constructs a new Darkpool instance.
-     * Initializes the GUI, storage, and task list.
-     * If loading data from storage fails, an empty task list is created.
-     */
     public Darkpool() {
         gui = new Gui();
         String filePath = "data/tasks.txt";
@@ -36,14 +26,6 @@ public class Darkpool {
         }
     }
 
-    /**
-     * Processes the user input and returns the response.
-     * Parses the input to create a command, executes the command, and returns the response.
-     * If the command indicates an exit, returns the goodbye message from the GUI.
-     *
-     * @param input The user input to be processed.
-     * @return The response after processing the input.
-     */
     public String getResponse(String input) {
         try {
             Command command = Parser.parse(input);
