+++ src/main/java/darkpool/parser/FindParser.java
@@ -4,8 +4,6 @@ import darkpool.command.Command;
 import darkpool.command.FindCommand;
 import darkpool.util.DarkpoolException;
 
-import java.util.Objects;
-
 public class FindParser {
 
     static public Command parse(String[] userInput) throws DarkpoolException {
