+++ src/main/java/darkpool/tasklist/ToFileString.java
@@ -1,0 +1,17 @@
+package darkpool.tasklist;
+
+import darkpool.task.Task;
+
+import java.util.ArrayList;
+
+public class ToFileString {
+    static public String toFileString(ArrayList<Task> taskList) {
+        StringBuilder fileString = new StringBuilder();
+
+        for (Task task : taskList) {
+            fileString.append(task.toFileString());
+        }
+
+        return String.valueOf(fileString);
+    }
+}
