+++ src/main/java/darkpool/util/Storage.java
@@ -10,31 +10,17 @@ import darkpool.task.Deadline;
 import darkpool.task.Event;
 import darkpool.task.Task;
 import darkpool.task.Todo;
+import darkpool.tasklist.TaskList;
 
-
-/**
- * Handles the loading and saving of tasks to and from a file.
- */
 
 public class Storage {
 
     private final String filePath;
 
-    /**
-     * Constructs a Storage object with the specified file path.
-     *
-     * @param filePath The path to the file where tasks are stored.
-     */
     public Storage(String filePath) {
         this.filePath = filePath;
     }
 
-    /**
-     * Checks if the file and its directory exist, and creates them if they do not.
-     *
-     * @return The file object representing the data file.
-     * @throws DarkpoolException If an I/O error occurs.
-     */
     private File validateData() throws DarkpoolException {
         File dataFile = new File(filePath);
 
@@ -50,12 +36,6 @@ public class Storage {
         return dataFile;
     }
 
-    /**
-     * Loads tasks from the data file.
-     *
-     * @return An ArrayList of tasks loaded from the file.
-     * @throws DarkpoolException If the file is not found or an error occurs while reading the file.
-     */
     public ArrayList<Task> loadData() throws DarkpoolException {
         ArrayList<Task> taskList = new ArrayList<>();
         String curTask;
@@ -77,12 +57,6 @@ public class Storage {
         return taskList;
     }
 
-    /**
-     * Saves tasks to the data file.
-     *
-     * @param taskList The list of tasks to be saved.
-     * @throws DarkpoolException If an I/O error occurs while writing to the file.
-     */
     public void saveData(TaskList taskList) throws DarkpoolException {
         try {
             java.io.FileWriter fileWriter = new java.io.FileWriter(filePath);
@@ -93,13 +67,6 @@ public class Storage {
         }
     }
 
-    /**
-     * Parses a task from a string representation.
-     *
-     * @param task The string representation of the task.
-     * @return The Task object parsed from the string.
-     * @throws DarkpoolException If the task string is invalid.
-     */
     private Task parseTask(String task) throws DarkpoolException {
         String[] taskParts = task.split(" \\| ");
         String type = taskParts[0];
